move collab ticket message to firestore
